Upgrade NextJS to v15 and other relevant packages
