+++ src/app/template/[resumeId]/TemplateSelectionPage.tsx
@@ -9,26 +9,27 @@ import {
   DialogContent,
   Grid,
 } from "@mui/material";
-import { Document, Font, Page, Text, PDFViewer } from "@react-pdf/renderer";
 
-import ResumeTemplate1 from "../../../components/renderer/ResumeTempate1";
-import ResumeTemplate2 from "../../../components/renderer/ResumeTemplate2";
-import ResumeTemplate3 from "../../../components/renderer/ResumeTemplate3";
 import { type Resume } from "../../resume/Resume";
 import { updateTemplateIdOfResume } from "@/service/api";
 import toast from "react-hot-toast";
+// import TemplateOptions from "./TemplateOptions";
+
+import dynamic from "next/dynamic";
+
+const TemplateOptions = dynamic(
+  () => import("./TemplateOptions").then((mod) => mod),
+  {
+    ssr: false,
+    loading: () => <p>Loading...</p>,
+  }
+);
 
 const templates = [
   { id: 1, image: "../img/template1.jpg" },
   { id: 2, image: "../img/template2.jpg" },
   { id: 3, image: "../img/template3.jpg" },
 ];
-
-// Ensure Lato font is loaded for consistency
-Font.register({
-  family: "Lato",
-  src: `https://fonts.gstatic.com/s/lato/v16/S6uyw4BMUTPHjx4wWw.ttf`,
-});
 
 interface TemplateSelectionPageProps {
   resume: Resume;
@@ -97,7 +98,7 @@ const TemplateSelectionPage: React.FC<TemplateSelectionPageProps> = ({
         ))}
       </Grid>
 
-      <TemplateRenderer resumeData={resume} templateType={templateType} />
+      <TemplateOptions resumeData={resume} templateType={templateType} />
       <Button
         className="button"
         variant="contained"
@@ -107,35 +108,6 @@ const TemplateSelectionPage: React.FC<TemplateSelectionPageProps> = ({
         Save
       </Button>
     </Container>
-  );
-};
-
-interface TemplateRendererProps {
-  resumeData: Resume;
-  templateType: number;
-}
-
-export const TemplateRenderer: React.FC<TemplateRendererProps> = ({
-  resumeData,
-  templateType,
-}) => {
-  if (templateType === 1) {
-    return <ResumeTemplate1 resume={resumeData} orientation="portrait" />;
-  }
-  if (templateType === 2) {
-    return <ResumeTemplate2 resume={resumeData} />;
-  }
-  if (templateType === 3) {
-    return <ResumeTemplate3 resume={resumeData} />;
-  }
-  return (
-    <PDFViewer width="100%" height="600">
-      <Document keywords="resume, ATS, multinational company" title="Resume">
-        <Page>
-          <Text>select the template to generate your resume PDF.</Text>
-        </Page>
-      </Document>
-    </PDFViewer>
   );
 };
